Move all store test utils into memory_store module
