+++ p2panda-rs/src/storage_provider/mod.rs
@@ -1,8 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 //! Storage provider traits needed for implementing custom p2panda storage solutions.
-mod domain;
 pub mod error;
 pub mod traits;
 pub mod utils;
-mod validation;
