+++ p2panda-rs/src/test_utils/memory_store/mod.rs
@@ -3,9 +3,12 @@
 //! Implementations of all `StorageProvider` traits.
 //!
 //! Used in the mock node and for testing.
+pub mod domain;
+pub mod helpers;
 mod provider;
 mod stores;
 mod types;
+pub mod validation;
 
 pub use provider::MemoryStore;
 pub use types::{EntryArgsResponse, PublishEntryResponse, StorageEntry, StorageLog};
